feat: options for Select
Add options for the Select input method. Also split up the input
functions into separate files.
